content/base/test/test_bug372964-2.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=372964
-->
<head>
  <title>Test for Bug 372964</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=372964">Mozilla Bug 372964</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 372964 **/

var eventCount = 0;

function listener(evt) {
  ++eventCount;
}

function runTest() {
  var ifr = document.getElementsByTagName("iframe")[0];
  var xhr = new ifr.contentWindow.XMLHttpRequest();
  xhr.addEventListener("foo", listener, false);
  var event = ifr.contentDocument.createEvent("Events");
  event.initEvent("foo", true, true);
  xhr.dispatchEvent(event);
  is(eventCount, 1, "Should have handled an event");
  ifr.contentDocument.open();
  ifr.contentDocument.close();
  event = ifr.contentDocument.createEvent("Events");
  event.initEvent("foo", true, true);
  xhr.dispatchEvent(event);
  is(eventCount, 1,
     "Shouldn't have handled an event because the context has changed");
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(runTest);

</script>
</pre>
<iframe>
</body>
</html>
